Feat: App.js 생성 및 NotionPage 코드 분리
