+++ src/index.js
@@ -1,8 +1,8 @@
-import NotionPage from './pages/NotionPage'
+import App from './components/App'
 
 const $target = document.querySelector('#app')
 
-new NotionPage({
+new App({
   $target,
   initialState: {
     documents: [],
