+++ src/pages/NotionPage.js
@@ -1,30 +1,19 @@
-import { createDocument, deleteSpecificDocument, getRootDocument, getSpecificDocument } from '../api/api'
+import { createDocument, deleteSpecificDocument } from '../api/api'
 import Content from '../components/Content'
 import DocumentsList from '../components/DocumentsList'
 
-export default function NotionPage({ $target, initialState }) {
+export default function NotionPage({
+  $target,
+  initialState,
+  fetchContentData,
+  fetchDocumentsData, 
+}) {
   const $notionPage = document.createElement('div')
   $target.appendChild($notionPage)
 
-  // documents 리스트 state
-  this.documents = initialState.documents
-
-  this.setDocuments = (nextDocuments) => {
-    this.documents = nextDocuments
-    documentsList.setState(nextDocuments)
-  }
-
-  // 편집기 state
-  this.editorContent = initialState.editorContent
-
-  this.setEditorContent = (nextContent) => {
-    this.editorContent = nextContent
-    content.setState(nextContent)
-  }
-
-  const documentsList = new DocumentsList({ 
+  this.documentsList = new DocumentsList({ 
     $target: $notionPage, 
-    initialState: this.documents,
+    initialState: initialState.documents,
     onClick: (id) => {
       fetchContentData(id)
     },
@@ -33,7 +22,6 @@ export default function NotionPage({ $target, initialState }) {
         title: "제목 없음",
         parent: id ? id : null
       }
-
       const newDocument = await createDocument(data)
 
       fetchDocumentsData()
@@ -41,32 +29,13 @@ export default function NotionPage({ $target, initialState }) {
     },
   })
 
-  const content = new Content({
+  this.content = new Content({
     $target: $notionPage,
-    initialState: this.editorContent,
+    initialState: initialState.editorContent,
     onDelete: async (id) => {
       await deleteSpecificDocument(id)
       fetchDocumentsData()
       fetchContentData()
-    }
+    },
   })
-
-  // 함수 파일 작성하기
-  // 전체 documentsList 불러오기
-  const fetchDocumentsData = async () => {
-    const documents = await getRootDocument()
-    this.setDocuments(documents)
-  }
-
-  // 특정 document의 content 불러오기
-  const fetchContentData = async (id = null) => {
-    if (id) {
-      const content = await getSpecificDocument(id)
-      this.setEditorContent(content)
-    } else {
-      this.setEditorContent(null)
-    }
-  }
-
-  fetchDocumentsData()
 }
